site.conf: update autoupdater branch names
testing is what was previously experimental, builds from gluons
master branch, where the development of the next major stable
release happens. It uses the latest stable branch of LEDE/OpenWrt.

unstable is the branch that relentlessly throws that work against
the OpenWrt master branch, allowing faster integration and testing
of new devices at the cost of a lacking quality assurance and
some degree of brokenness here and there.